Rename to applications.ts
