+++ delete-pull-request-namespaces/src/applications.ts
